+++ src/lib.rs
@@ -138,6 +138,7 @@ pub mod noisegate;
 pub mod processing;
 #[cfg(feature = "pulse-backend")]
 pub mod pulsedevice;
+pub mod race;
 #[cfg(feature = "websocket")]
 pub mod socketserver;
 pub mod statefile;