+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## v4.0.0
 New features:
+- Add RACE processor.
 - Add option for custom log filtering.
+Changes:
+- Add microsecond delay unit.
 
 ## v3.0.0
 New features: